Tidy up serializable extensions
