+++ lib/jsonapi_compliable.rb
@@ -10,47 +10,9 @@ require 'dry-types'
 require 'jsonapi_errorable'
 
 require 'jsonapi/serializable'
-# Temporary fix until fixed upstream
-# https://github.com/jsonapi-rb/jsonapi-serializable/pull/102
-JSONAPI::Serializable::Resource.class_eval do
-  def requested_relationships(fields)
-    @_relationships
-  end
-end
-
-# This library looks up a serializer based on the record's class name
-# This wouldn't work for us, since a model may be associated with
-# multiple resources.
-# Instead, this variable is assigned when the query is resolved
-# To ensure we always render with the *resource* serializer
-JSONAPI::Serializable::Renderer.class_eval do
-  def _build(object, exposures, klass)
-    klass = object.instance_variable_get(:@__serializer_klass)
-    klass.new(exposures.merge(object: object))
-  end
-end
-
-# See above comment
-JSONAPI::Serializable::Relationship.class_eval do
-  def data
-    @_resources_block = proc do
-      resources = yield
-      if resources.nil?
-        nil
-      elsif resources.respond_to?(:to_ary)
-        Array(resources).map do |obj|
-          klass = obj.instance_variable_get(:@__serializer_klass)
-          klass.new(@_exposures.merge(object: obj))
-        end
-      else
-        klass = resources.instance_variable_get(:@__serializer_klass)
-        klass.new(@_exposures.merge(object: resources))
-      end
-    end
-  end
-end
 
 require "jsonapi_compliable/version"
+require "jsonapi_compliable/jsonapi_serializable_ext"
 require "jsonapi_compliable/configuration"
 require "jsonapi_compliable/context"
 require "jsonapi_compliable/errors"
